Move PSBT unit test to separate file
